remove round num in LNOT module. Now COT and NOT support num > updateNum.

+++ mpc4j-s2pc-pcg/src/main/java/edu/alibaba/mpc4j/s2pc/pcg/ot/cot/AbstractCotReceiver.java
@@ -20,7 +20,7 @@ public abstract class AbstractCotReceiver extends AbstractTwoPartyPto implements
     /**
      * update num
      */
-    protected long updateNum;
+    protected int updateNum;
     /**
      * choices
      */
