+++ mpc4j-s2pc-pcg/src/main/java/edu/alibaba/mpc4j/s2pc/pcg/ot/cot/AbstractCotSender.java
@@ -26,7 +26,7 @@ public abstract class AbstractCotSender extends AbstractTwoPartyPto implements C
     /**
      * update num
      */
-    protected long updateNum;
+    protected int updateNum;
     /**
      * num
      */
